7/23 23：58
修改 lcd 驱动        还原

+++ Inc/lcd.h
@@ -30,17 +30,13 @@ typedef struct
 	u16 setycmd;		//设置y坐标指令 
 }_lcd_dev; 	  
 
-////LCD参数
-//extern _lcd_dev lcddev;	//管理LCD重要参数
-////LCD的画笔颜色和背景色	   
-//extern u32  POINT_COLOR;//默认红色    
-//extern u32  BACK_COLOR; //背景颜色.默认为白色
-
 //LCD参数
 extern _lcd_dev lcddev;	//管理LCD重要参数
 //LCD的画笔颜色和背景色	   
-extern u16  POINT_COLOR;//默认红色    
-extern u16  BACK_COLOR; //背景颜色.默认为白色
+extern u32  POINT_COLOR;//默认红色    
+extern u32  BACK_COLOR; //背景颜色.默认为白色
+
+
 
 //////////////////////////////////////////////////////////////////////////////////	 
 //-----------------MCU屏 LCD端口定义---------------- 
@@ -102,15 +98,15 @@ typedef struct
 void LCD_Init(void);													   	//初始化
 void LCD_DisplayOn(void);													//开显示
 void LCD_DisplayOff(void);													//关显示
-void LCD_Clear(u16 Color);	 												//清屏
+void LCD_Clear(u32 Color);	 												//清屏
 void LCD_SetCursor(u16 Xpos, u16 Ypos);										//设置光标
 void LCD_DrawPoint(u16 x,u16 y);											//画点
-void LCD_Fast_DrawPoint(u16 x,u16 y,u16 color);								//快速画点
-u16  LCD_ReadPoint(u16 x,u16 y); 											//读点 
+void LCD_Fast_DrawPoint(u16 x,u16 y,u32 color);								//快速画点
+u32  LCD_ReadPoint(u16 x,u16 y); 											//读点 
 void LCD_Draw_Circle(u16 x0,u16 y0,u8 r);						 			//画圆
 void LCD_DrawLine(u16 x1, u16 y1, u16 x2, u16 y2);							//画线
 void LCD_DrawRectangle(u16 x1, u16 y1, u16 x2, u16 y2);		   				//画矩形
-void LCD_Fill(u16 sx,u16 sy,u16 ex,u16 ey,u16 color);		   				//填充单色
+void LCD_Fill(u16 sx,u16 sy,u16 ex,u16 ey,u32 color);		   				//填充单色
 void LCD_Color_Fill(u16 sx,u16 sy,u16 ex,u16 ey,u16 *color);				//填充指定颜色
 void LCD_ShowChar(u16 x,u16 y,u8 num,u8 size,u8 mode);						//显示一个字符
 void LCD_ShowNum(u16 x,u16 y,u32 num,u8 len,u8 size);  						//显示一个数字
